feature: rename constants to mocks
