+++ src/components/header.tsx
@@ -4,7 +4,7 @@ import { Dispatch, SetStateAction, useState } from 'react'
 import { Dialog } from '@headlessui/react'
 import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline'
 
-import { navigation } from '@/constants/header'
+import { navigation } from '@/mocks/header'
 
 const Header = () => {
   const [mobileMenuOpen, setMobileMenuOpen] = useState<boolean>(false)
