+++ src/mocks/footer.ts
